- form review  - removed all form classes in "/Forms" and merged them with livewire form class - added $buttonType for form actions - default register will set default channel to email - cancel-account button for UserProfile - UserProfile and User DataTable permission checks for delete/cancel-account

+++ app/Http/Livewire/DataTable/User.php
@@ -243,6 +243,11 @@ class User extends BaseDataTable
             return false;
         }
 
+        if (!$this->canManage()) {
+            $this->addErrorMessage('Permission denied');
+            return false;
+        }
+
         /** @var UserModel $user */
         if ($user = app(\App\Models\User::class)->with([])->find($itemId)) {
             $result = $user->deleteIn3Steps();
